CR* should be 8 bytes on x64 platform.

+++ Memory_Address_Mapping/sys_reg.c
@@ -15,7 +15,7 @@
   }__attribute__((packed));
 */
 
- static unsigned int cr0,cr3,cr4;
+ static unsigned long cr0,cr3,cr4;
  static struct desc_ptr gdtr;
 
  static int my_proc_show(struct seq_file *m, void *v)
